Moved UserDomainValidation to FMS/Helpers. Added Unit Test for UserDomainValidation
